created Exceptions

+++ src/main/java/org/example/homemade2/service/FavouritePaymentService.java
@@ -7,11 +7,14 @@ import org.example.homemade2.dto.GetResponseFavPayDTO;
 import org.example.homemade2.dto.ResponseFavPayDTO;
 import org.example.homemade2.entity.FavouritePayment;
 import org.example.homemade2.entity.HomeDetails;
+import org.example.homemade2.exceptions.AlreadyExistException;
+import org.example.homemade2.exceptions.NotFoundException;
 import org.example.homemade2.mapper.FavouritePaymentMapper;
 import org.example.homemade2.repository.FavouritePaymentRepository;
 import org.example.homemade2.repository.HomeDetailsRepository;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -24,24 +27,21 @@ public class FavouritePaymentService {
         Optional<HomeDetails> homeDetails = homeDetailsRepository.findHomeDetailsById(createFavouritePayment.homeDetails_id());
         ResponseFavPayDTO responseFavPayDTO = new ResponseFavPayDTO();
         if (homeDetails.isPresent()) {
-            try {
-                FavouritePayment favouritePayment = FavouritePaymentMapper.toEntity(createFavouritePayment);
+            FavouritePayment favouritePayment = FavouritePaymentMapper.toEntity(createFavouritePayment);
+            List<FavouritePayment> favouritePayments = favouritePaymentRepository.findFavouritePaymentByHomeDetails(homeDetails.get());
+            if (favouritePayments.isEmpty()) {
                 favouritePayment.setHomeDetails(homeDetails.get());
                 favouritePaymentRepository.save(favouritePayment);
                 responseFavPayDTO.setIsSuccess(true);
                 responseFavPayDTO.setMessage("Favourite payment successfully created");
                 log.info("Favourite payment successfully created : {}", favouritePayment.getId());
-            }catch (Exception e) {
-                responseFavPayDTO.setIsSuccess(false);
-                responseFavPayDTO.setMessage("Favourite payment failed");
-                log.error("Favourite payment failed : {}", e.getMessage());
+            }else {
+                log.error("Favourite payment already exist");
+                throw new AlreadyExistException("Favourite payment already exist");
             }
-
-
         }else {
-            responseFavPayDTO.setIsSuccess(false);
-            responseFavPayDTO.setMessage("Home details not found");
-            log.error("Home details not found : {}", responseFavPayDTO.getIsSuccess());
+            log.error("Home details not found");
+            throw new NotFoundException("Home details not found");
         }
         return responseFavPayDTO;
     }
@@ -51,53 +51,46 @@ public class FavouritePaymentService {
         Optional<HomeDetails> homeDetails = homeDetailsRepository.findHomeDetailsById(createFavouritePayment.homeDetails_id());
         ResponseFavPayDTO responseFavPayDTO = new ResponseFavPayDTO();
         if (homeDetails.isPresent()) {
-            try {
                 FavouritePayment favouritePayment = FavouritePaymentMapper.toEntity(createFavouritePayment);
                 favouritePayment.setHomeDetails(homeDetails.get());
                 favouritePaymentRepository.save(favouritePayment);
                 responseFavPayDTO.setIsSuccess(true);
                 responseFavPayDTO.setMessage("Favourite payment successfully updated");
                 log.info("Favourite payment successfully updated");
-
-
-            }catch (Exception e) {
-                responseFavPayDTO.setIsSuccess(false);
-                responseFavPayDTO.setMessage("Favourite payment failed");
-                log.error("Favourite payment fail : {}", e.getMessage());
-            }
         }else {
-            responseFavPayDTO.setIsSuccess(false);
-            responseFavPayDTO.setMessage("Home details not found");
-            log.error("Home details not found: {}", responseFavPayDTO.getIsSuccess());
+            log.error("Home details not found ");
+            throw new NotFoundException("Home details not found");
         }
         return responseFavPayDTO;
     }
 
     public ResponseFavPayDTO deleteFavouritePaymentById(Long id){
         ResponseFavPayDTO responseFavPayDTO = new ResponseFavPayDTO();
-        try {
+        Optional<FavouritePayment> favouritePayment = favouritePaymentRepository.findById(id);
+        if (favouritePayment.isPresent()) {
             favouritePaymentRepository.deleteById(id);
             responseFavPayDTO.setIsSuccess(true);
             responseFavPayDTO.setMessage("Favourite payment successfully deleted");
             log.info("Favourite payment successfully deleted");
-
-        }catch (Exception e){
-            responseFavPayDTO.setIsSuccess(false);
-            responseFavPayDTO.setMessage("Favourite payment failed");
-            log.error("Favourite payment failed: {}", e.getMessage());
+        }else {
+            log.error("Favourite payment not found");
+            throw new NotFoundException("Favourite payment not found");
         }
+
         return responseFavPayDTO;
     }
 
     public GetResponseFavPayDTO findFavouritePaymentById(Long id){
         GetResponseFavPayDTO responseFavPayDTO = new GetResponseFavPayDTO();
-        try {
-            FavouritePayment favouritePayment = favouritePaymentRepository.findById(id).get();
-            responseFavPayDTO.setHomeDetails(favouritePayment.getHomeDetails());
-            responseFavPayDTO.setAmount(favouritePayment.getAmount());
+        Optional<FavouritePayment> favouritePayment = favouritePaymentRepository.findById(id);
+        if (favouritePayment.isPresent()) {
+            FavouritePayment favouritePayment2 = favouritePayment.get();
+            responseFavPayDTO.setHomeDetails(favouritePayment2.getHomeDetails());
+            responseFavPayDTO.setAmount(favouritePayment2.getAmount());
             log.info("Favourite payment successfully retrieved");
-        }catch (Exception e){
-            log.error("Favourite payment retrieving failed : {}", e.getMessage());
+        }else {
+            log.error("Favourite payment not found ");
+            throw new NotFoundException("Favourite payment not found");
         }
         return responseFavPayDTO;
     }
